--- mlir/include/mlir/Dialect/AMX/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_mlir_dialect(AMX amx)
-add_mlir_doc(AMX AMX Dialects/ -gen-dialect-doc -dialect=amx)
-
-add_mlir_interface(AMXInterfaces)
-add_dependencies(MLIRAMXIncGen MLIRAMXInterfacesIncGen)
