--- mlir/lib/Dialect/AMX/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(IR)
-add_subdirectory(Transforms)
