--- mlir/lib/Dialect/AMX/Transforms/CMakeLists.txt
@@ -1,9 +1,0 @@
-add_mlir_dialect_library(MLIRAMXTransforms
-  LegalizeForLLVMExport.cpp
-
-  LINK_LIBS PUBLIC
-  MLIRAMXDialect
-  MLIRIR
-  MLIRLLVMCommonConversion
-  MLIRLLVMDialect
-  )
