+++ mlir/test/Conversion/VectorToLLVM/pass-option-serialization.mlir
@@ -18,7 +18,6 @@
 
 // CHECK: builtin.module(
 // CHECK-SAME: convert-vector-to-llvm{
-// CHECK-SAME: enable-amx={{[aA-zZ0-9]+}}
 // CHECK-SAME: enable-arm-neon={{[aA-zZ0-9]+}}
 // CHECK-SAME: enable-arm-sve={{[aA-zZ0-9]+}}
 // CHECK-SAME: enable-x86={{[aA-zZ0-9]+}}
