+++ mlir/test/Integration/Dialect/Vector/CPU/X86/AMX/lit.local.cfg
@@ -1,7 +1,7 @@
 import sys
 
 # AMX tests must be enabled via build flag.
-if not config.mlir_run_amx_tests:
+if not config.mlir_run_x86_amx_tests:
     config.unsupported = True
 
 # No JIT on win32.
