+++ mlir/test/Integration/Dialect/Vector/CPU/X86/dot.mlir
@@ -5,14 +5,15 @@
 
 func.func @entry() -> i32 {
   %i0 = arith.constant 0 : i32
-  %i4 = arith.constant 4 : i32
+  %c0 = arith.constant 0 : index
+  %c4 = arith.constant 4 : index
 
   %a = arith.constant dense<[1.0,  2.0,  3.0,  4.0,  5.0,  6.0,  7.0,  8.0]> : vector<8xf32>
   %b = arith.constant dense<[9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0]> : vector<8xf32>
   %r = x86.avx.intr.dot %a, %b : vector<8xf32>
 
-  %1 = vector.extract %r[%i0] : f32 from vector<8xf32>
-  %2 = vector.extract %r[%i4] : f32 from vector<8xf32>
+  %1 = vector.extract %r[%c0] : f32 from vector<8xf32>
+  %2 = vector.extract %r[%c4] : f32 from vector<8xf32>
   %d = arith.addf %1, %2 : f32
 
   // CHECK: ( 110, 110, 110, 110, 382, 382, 382, 382 )
